+++ taplo-ide/vscode/src/commands/tomlToJson.ts
@@ -34,32 +34,94 @@ export function register(
             params
           );
 
-          if (res.errors?.length ?? 0 !== 0) {
-            for (const err of res.errors!) {
-              const out = getOutput();
-              out.appendLine(
-                `Selection Parse Error: (${editor.document.fileName}) ${err}`
-              );
-            }
+          const out = getOutput();
 
-            await vscode.window.showErrorMessage(
-              "Copy failed! (Details under the output tab)"
+          if (res.errors?.length ?? 0 !== 0) {
+            let errLines = `Selection Parse Errors (${editor.document.fileName}):`;
+            for (const err of res.errors!) {
+              errLines += `\n\t${err}`;
+            }
+            out.appendLine(errLines);
+
+            const show = await vscode.window.showErrorMessage(
+              "Copying has failed!",
+              "Show Details"
             );
 
+            if (show) {
+              out.show();
+            }
             return;
           }
 
           try {
             await clipboardy.write(res.text!);
           } catch (e) {
-            getOutput().appendLine(e);
-            await vscode.window.showErrorMessage(
-              "Copy failed! (Details under the output tab)"
+            out.appendLine(`Couldn't write to clipboard: ${e}`);
+            const show = await vscode.window.showErrorMessage(
+              "Copying has failed!",
+              "Show Details"
             );
+
+            if (show) {
+              out.show();
+            }
             return;
           }
 
           await vscode.window.showInformationMessage("JSON has been copied!");
+        }
+      ),
+      vscode.commands.registerTextEditorCommand(
+        "evenBetterToml.pasteTomlAsJson",
+        async (editor) => {
+          const out = getOutput();
+          let input;
+          try {
+            input = await clipboardy.read();
+          } catch (e) {
+            out.appendLine(`Failed to read from clipboard:${e}`);
+            const show = await vscode.window.showErrorMessage(
+              "Paste from clipboard has failed!",
+              "Show Details"
+            );
+
+            if (show) {
+              out.show();
+            }
+            return;
+          }
+
+          let params: requestExt.TomlToJson.Params = {
+            text: input,
+          };
+
+          const res = await c.sendRequest<requestExt.TomlToJson.Response>(
+            requestExt.TomlToJson.METHOD,
+            params
+          );
+
+          if (res.errors?.length ?? 0 !== 0) {
+            let errLines = `Clipboard Parse Errors:`;
+            for (const err of res.errors!) {
+              errLines += `\n\t${err}`;
+            }
+            out.appendLine(errLines);
+
+            const show = await vscode.window.showErrorMessage(
+              "Paste from clipboard has failed!",
+              "Show Details"
+            );
+
+            if (show) {
+              out.show();
+            }
+            return;
+          }
+
+          editor.edit((e) => {
+            e.replace(editor.selection, res.text!);
+          });
         }
       )
     );